requirements.txt
python-dotenv
langchain
langchain_community
langchain_openai
streamlit
streamlit_chat
databricks-sql-connector
databricks-sqlalchemy
duckdb
kagglehub
duckdb-engine